+++ src/Database/Adapter/MysqlAdapter.php
@@ -29,6 +29,16 @@ class MysqlAdapter extends PdoAdapter
         return $this->queryBuilder;
     }
 
+    public function buildDoNotCheckForeignKeysQuery(): string
+    {
+        return 'SET FOREIGN_KEY_CHECKS = 0;';
+    }
+
+    public function buildCheckForeignKeysQuery(): string
+    {
+        return 'SET FOREIGN_KEY_CHECKS = 1;';
+    }
+
     protected function loadDatabase(): string
     {
         /** @var string $currentDatabase */